# openapi.yaml
openapi: 3.0.1
info:
  title: Customer AS2 Connections
  description: Retrieve customer AS2 connections based on specified parameters.
  version: '1.0'
  x-ibm-application-name: Customer AS2 Connections
  x-ibm-application-id: customer-as2-connections
  x-ibm-annotations: 'true'
  x-ibm-skill-type: imported
servers:
  - url: https://certapi.onrender.com
security:
  - dummyAuth: []
paths:
  /internal/customer_as2_connections:
    get:
      summary: Retrieve customer AS2 connections
      description: Retrieve customer AS2 connections based on as2UrlDestination and/or as2IdTp parameters.
      parameters:
        - name: as2UrlDestination
          in: query
          required: false
          schema:
            type: string
          description: The AS2 URL destination to filter the connections.
        - name: as2IdTp
          in: query
          required: false
          schema:
            type: string
          description: The AS2 ID TP to filter the connections.
      responses:
        '200':
          description: Successful retrieval of the AS2 connection details.
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    Company_Name:
                      type: string
                      example: Acme Corp
                    Relationship_Type:
                      type: string
                      example: Partner
                    Pod_Name:
                      type: string
                      example: PodA
                    Business_Alias:
                      type: string
                      example: AcmeAlias
                    AS2ID_Customer:
                      type: string
                      example: AS2ID123
                    Receiver:
                      type: string
                      example: ReceiverName
                    Decryption_Customer:
                      type: string
                      example: DecryptionCert
                    Decryption_Exp_C:
                      type: string
                      example: 2024-12-31
                    Signing_Cert_Customer:
                      type: string
                      example: SigningCert
                    Signing_Exp_C:
                      type: string
                      example: 2024-12-31
                    SSL_Client_Auth_Key_C:
                      type: string
                      example: SSLKey
                    SSL_Key_Exp_C:
                      type: string
                      example: 2024-12-31
                    Remote_Business_Alias:
                      type: string
                      example: RemoteAlias
                    AS2_ID_TP:
                      type: string
                      example: AS2IDTP123
                    Destination:
                      type: string
                      example: DestinationURL
                    Encryption_Cert_TP:
                      type: string
                      example: EncryptionCert
                    EncryptionExp:
                      type: string
                      example: 2024-12-31
                    Sign_Valid_Cert_TP:
                      type: string
                      example: SignCert
                    Sign_Valid_Exp_TP:
                      type: string
                      example: 2024-12-31
                    Description:
                      type: string
                      example: Connection details
                    Protocol_Profile_ID:
                      type: string
                      example: ProtocolID123
                    Environment_Loc:
                      type: string
                      example: Production
        '400':
          description: Missing required parameters.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: No connection details found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  schemas:
    ErrorResponse:
      type: object
      properties:
        message:
          type: string
          example: Connection details not found!
  securitySchemes:
    dummyAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT